Dead code cleanup

+++ js/rhino/src/org/mozilla/javascript/Decompiler.java
@@ -862,7 +862,7 @@ public class Decompiler
                 lbits = (long)source.charAt(offset) << 48;
                 lbits |= (long)source.charAt(offset + 1) << 32;
                 lbits |= (long)source.charAt(offset + 2) << 16;
-                lbits |= (long)source.charAt(offset + 3);
+                lbits |= source.charAt(offset + 3);
                 if (type == 'J') {
                     number = lbits;
                 } else {
