+++ js/rhino/src/org/mozilla/javascript/WrapFactory.java
@@ -83,7 +83,7 @@ public class WrapFactory
             if (staticType == Void.TYPE)
                 return Undefined.instance;
             if (staticType == Character.TYPE)
-                return new Integer((int) ((Character) obj).charValue());
+                return new Integer(((Character) obj).charValue());
             return obj;
         }
         if (!isJavaPrimitiveWrap()) {
